+++ aerich/cli.py
@@ -10,12 +10,11 @@ from click import Context, UsageError
 from tomlkit.exceptions import NonExistentKey
 from tortoise import Tortoise
 
+from aerich import Command
+from aerich.enums import Color
 from aerich.exceptions import DowngradeError
 from aerich.utils import add_src_path, get_tortoise_config
-
-from . import Command
-from .enums import Color
-from .version import __version__
+from aerich.version import __version__
 
 CONFIG_DEFAULT_VALUES = {
     "src_folder": ".",
@@ -27,11 +26,11 @@ def coro(f):
     def wrapper(*args, **kwargs):
         loop = asyncio.get_event_loop()
 
-        # Close db connections at the end of all all but the cli group function
+        # Close db connections at the end of all but the cli group function
         try:
             loop.run_until_complete(f(*args, **kwargs))
         finally:
-            if f.__name__ != "cli":
+            if f.__name__ not in ["cli", "init_db"]:
                 loop.run_until_complete(Tortoise.close_connections())
 
     return wrapper
@@ -55,10 +54,10 @@ async def cli(ctx: Context, config, app):
 
     invoked_subcommand = ctx.invoked_subcommand
     if invoked_subcommand != "init":
-        if not Path(config).exists():
+        config_path = Path(config)
+        if not config_path.exists():
             raise UsageError("You must exec init first", ctx=ctx)
-        with open(config, "r") as f:
-            content = f.read()
+        content = config_path.read_text()
         doc = tomlkit.parse(content)
         try:
             tool = doc["tool"]["aerich"]
@@ -193,18 +192,19 @@ async def init(ctx: Context, tortoise_orm, location, src_folder):
     # check that we can find the configuration, if not we can fail before the config file gets created
     add_src_path(src_folder)
     get_tortoise_config(ctx, tortoise_orm)
-
-    with open(config_file, "r") as f:
-        content = f.read()
-    doc = tomlkit.parse(content)
+    config_path = Path(config_file)
+    if config_path.exists():
+        content = config_path.read_text()
+        doc = tomlkit.parse(content)
+    else:
+        doc = tomlkit.parse("[tool.aerich]")
     table = tomlkit.table()
     table["tortoise_orm"] = tortoise_orm
     table["location"] = location
     table["src_folder"] = src_folder
     doc["tool"]["aerich"] = table
 
-    with open(config_file, "w") as f:
-        f.write(tomlkit.dumps(doc))
+    config_path.write_text(tomlkit.dumps(doc))
 
     Path(location).mkdir(parents=True, exist_ok=True)
 
@@ -214,15 +214,17 @@ async def init(ctx: Context, tortoise_orm, location, src_folder):
 
 @cli.command(help="Generate schema and generate app migrate location.")
 @click.option(
+    "-s",
     "--safe",
     type=bool,
+    is_flag=True,
     default=True,
     help="When set to true, creates the table only when it does not already exist.",
     show_default=True,
 )
 @click.pass_context
 @coro
-async def init_db(ctx: Context, safe):
+async def init_db(ctx: Context, safe: bool):
     command = ctx.obj["command"]
     app = command.app
     dirname = Path(command.location, app)
@@ -248,7 +250,8 @@ async def init_db(ctx: Context, safe):
 @coro
 async def inspectdb(ctx: Context, table: List[str]):
     command = ctx.obj["command"]
-    await command.inspectdb(table)
+    ret = await command.inspectdb(table)
+    click.secho(ret)
 
 
 def main():