+++ tests/models.py
@@ -29,6 +29,7 @@ class User(Model):
     is_active = fields.BooleanField(default=True, description="Is Active")
     is_superuser = fields.BooleanField(default=False, description="Is SuperUser")
     intro = fields.TextField(default="")
+    longitude = fields.DecimalField(max_digits=10, decimal_places=8)
 
 
 class Email(Model):